--- HW12WebService/src/main/java/ru/otus/core/service/DbServiceUser.java
@@ -1,13 +1,0 @@
-package ru.otus.core.service;
-
-import ru.otus.core.model.User;
-
-import java.util.Optional;
-
-public interface DbServiceUser {
-    long saveUser(User user);
-    Optional<User> getUserById(long id);
-    long saveOrUpdate(User user);
-    Optional<User> getUserByLogin (String login);
-    Optional<User> getRandomUser();
-}
