--- HW12WebService/src/main/java/ru/otus/webserviceHW/servlet/AuthorizationFilter.java
@@ -1,38 +1,0 @@
-package ru.otus.webserviceHW.servlet;
-
-
-import javax.servlet.*;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import java.io.IOException;
-
-public class AuthorizationFilter implements Filter {
-    private ServletContext context;
-
-    @Override
-    public void init(FilterConfig filterConfig) throws ServletException {
-        this.context=filterConfig.getServletContext();
-    }
-
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-        HttpServletRequest request = (HttpServletRequest) servletRequest;
-        HttpServletResponse response = (HttpServletResponse)servletResponse;
-
-        String uri = request.getRequestURI();
-        this.context.log("Requested Resource:"+uri);
-        HttpSession session = request.getSession(false);
-
-        if(session == null){
-            response.sendRedirect("/login");
-        }else{
-            filterChain.doFilter(servletRequest, servletResponse);
-        }
-    }
-
-    @Override
-    public void destroy() {
-
-    }
-}
