--- HW12WebService/src/main/java/ru/otus/webserviceHW/servlet/UsersApiServlet.java
@@ -1,39 +1,0 @@
-package ru.otus.webserviceHW.servlet;
-
-import com.google.gson.Gson;
-import ru.otus.core.dao.UserDao;
-import ru.otus.core.model.User;
-import ru.otus.core.service.DbServiceUser;
-
-import javax.servlet.ServletOutputStream;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class UsersApiServlet extends HttpServlet {
-    private static final int ID_PATH_PARAM_POSITION = 1;
-
-    private final DbServiceUser dbServiceUser;
-    private final Gson gson;
-
-    public UsersApiServlet(DbServiceUser dbServiceUser, Gson gson) {
-        this.dbServiceUser = dbServiceUser;
-        this.gson = gson;
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        User user = dbServiceUser.getUserById(extractIdFromRequest(request)).orElse(null);
-
-        response.setContentType("application/json;charset=UTF-8");
-        ServletOutputStream out = response.getOutputStream();
-        out.print(gson.toJson(user));
-    }
-
-    private long  extractIdFromRequest(HttpServletRequest request) {
-        String[] path = request.getPathInfo().split("/");
-        String id = (path.length>1)? path[ID_PATH_PARAM_POSITION]: String.valueOf(-1);
-        return Long.parseLong(id);
-    }
-}
