+++ HW13Spring/src/main/java/ru/otus/App.java
@@ -1,0 +1,14 @@
+package ru.otus;
+
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+
+@ComponentScan
+public class App {
+    public static void main(String[] args) {
+        ApplicationContext ctx = new AnnotationConfigApplicationContext(WebConfig.class);
+    }
+}
