+++ HW13Spring/src/main/java/ru/otus/core/dao/UserDaoException.java
