+++ HW13Spring/src/main/java/ru/otus/core/model/User.java
