+++ HW13Spring/src/main/java/ru/otus/core/service/DbServiceException.java
