+++ HW13Spring/src/main/java/ru/otus/core/sessionmanager/SessionManagerException.java
