+++ HW13Spring/src/main/java/ru/otus/hibernate/sessionmanager/DatabaseSessionHibernate.java
