+++ HW13Spring/src/main/java/ru/otus/spring/GreetingController.java
@@ -1,0 +1,8 @@
+package ru.otus.spring;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class GreetingController {
+
+}
